Breakout remaining gulp tasks
