+++ gulp/tasks/build.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task('build', ['buildJs', 'buildCss', 'buildImg']);
