+++ gulp/tasks/connect.js
@@ -1,0 +1,8 @@
+var gulp = require('gulp'),
+    connect = require('gulp-connect');
+
+
+gulp.task('connect', connect.server({
+  root: ['example'],
+  port: 4000
+}));
