+++ gulp/tasks/default.js
@@ -1,0 +1,5 @@
+var gulp = require('gulp');
+
+gulp.task('default', ['setup'], function(){
+  gulp.run('build');
+});
