+++ gulp/tasks/help.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task('help', help);
