+++ gulp/tasks/setup.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task('setup', ['updateBowerVersion', 'buildClean']);
