+++ gulp/tasks/tag.js
@@ -1,0 +1,5 @@
+var gulp = require('gulp');
+
+gulp.task('tag', function(){
+  git.tag('v' + version, 'Release v' + version);
+});
